// frontend/src/shared/api/generated/models/activityRoleTypeResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * CodigoActivo API
 * Backend API for CodigoActivo (cookie session auth + CSRF).
 * OpenAPI spec version: v1
 */

export interface ActivityRoleTypeResponse {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
}